Small file name refactor to fulfill PEP 8 naming conventions.
