+++ src/aoc2/cube_container.py
